refactor(#137/*): category success code global/response 로 이동
